refactor(scripts): move shared utilities into scripts/lib/
Moves 8 foundational utilities (shared-constants, intent-resolver,
resolve-unit, derive-scope, scope-aware-runner, screen-id,
snapshot-store, resolve-node.sh) into scripts/lib/. Updates require()
paths in 3 importing scripts (transform-to-hifi, fm-tree-to-flow-data,
validate-flow-data), 7+ importing tests, and string-path consumers in
CLAUDE.md, screen-generator agent, and 4 SKILL.md files. Fixes
__dirname depth in shared-constants.js (now scripts/lib/ → ../.. to
reach plugin root) and updates parity-check.md reference doc.
