+++ plugins/actian-design-system/scripts/lib/derive-scope.js
