+++ plugins/actian-design-system/scripts/lib/intent-resolver.js
