+++ plugins/actian-design-system/scripts/lib/resolve-node.sh
