+++ plugins/actian-design-system/scripts/lib/screen-id.js
