+++ plugins/actian-design-system/scripts/lib/shared-constants.js
@@ -21,6 +21,7 @@ function loadRegistry(name) {
     const filePath = path.join(
       __dirname,
       "..",
+      "..",
       "docs",
       "generated",
       name + ".json",
