+++ plugins/actian-design-system/scripts/lib/snapshot-store.js
